Create sentiment1
I wanted to check the sentiment score of the reviews. In order to do that I decided to use the VADER sentiment analysis tool because it checks also the intensity of each review using several parameters: punctuation, capitalization, degree modifiers, shift in polarity due to “but” and examining the tri-gram before a sentiment-laden lexical feature to catch polarity negation.

If we look at the combinations we can see that:
1. If the positive review is everything and the negative review is nothing or no negative - > the review is fully positive.
2. If the negative review is everything and the positive review is nothing or no positive - > the review is fully negative.
3. If the negative review is nothing or no negative we need to check the intensity of the positive review.
4. If the positive review is nothing or no positive we need to check the intensity of the negative review.
5. All the other reviews can be checked regularly. 

After I classified the reviews, I needed to analyze them base on the group they are in. 
for example, if the review count is 1 it means the review sentiment score will be 1, total positive and if the review count is 2 it means the review sentiment score will be -1, total negative. Furthermore, if the review count is 3 we need to analyze only the positive review and if the review count is 4 we need to analyze only the negative review. In reviews that has a 5 score we need to analyze both.

I linked the VADER sentiment intensity analyzer in order to use it for our sentiment analysis. I created a function that uses the review count classification and return 1 for label 1, -1 for label 2 and the compound sentiment score of each review that has labels 3,4 or 5.