Create sentiment2
See what the pattern of the reviews is based on their sentiment score and the reviewer score and if there’s a correlation between them.

The scores from both the reviewers and the sentiment analysis tends to be positive, although in the sentiment analysis there’s also a massive amount of neutral reviews score. 

We can see that the strongest focus is on the higher scores of each axis, and we also have a weaker focus on the neutral sentiment score. Most of the reviewers scores are between 8.5-10 and most of the sentiment scores are between 0.6-1. The Pearson’s correlation coefficient is 0.49, if the value lies between 0.30-0.49 it reflects a moderate degree of a medium correlation. However, our PCC is closer to the high degree which reflects strong correlation. It means that there is a positive connection between the reviewer score and the review sentiment score. 

I tried to see the pattern of the average scores based on the reviewer nationality.

We can see that the division is normal, most of the values are in the center of the division between 8-9 on the reviewer score axis and between 0.4-0.6 on the review sentiment score. The Pearson’s correlation coefficient is 0.47, which reflects again of a medium connection between those two.

The most interesting thing about this graph and the one before is that even though we can see some negative and neutral scores in the sentiment score the reviewers scores don’t go below 5, which means that even if the reviewer had a negative experience in the hotel, he provided a neutral score or even a high one.